+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.1.0] - 2018-05-31 
+<<<<<<< HEAD
 
+=======
+ 
+>>>>>>> master
 ### Added
 - CONTRIBUTE.md
 - *getting started* chapter in docs/