+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - *test_tapering.py* in nessi/signal/tests
 - *test_filtering.py* in nessi/signal/tests
 
+### Modified
+- README.md
+
 ### Fixed
 - remake html documentation
 - *signal.time_window* issue for 1D signal